perbaikan error delete contraint

+++ app/Http/Livewire/DataMaster/Spp/SppIndex.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Livewire\DataMaster\Spp;
 
+use App\Models\Siswa;
 use App\Models\Spp;
 use Livewire\Component;
 use App\Traits\ListenerTrait;
@@ -27,12 +28,19 @@ class SppIndex extends Component
 
     public function sppDelete($id)
     {
-        $spp = Spp::find($id);
-        if ($spp) {
-            $spp->delete();
-            $this->emit('toastify', ['success','Berhasil Menghapus Spp']);
+        $siswa = Siswa::where('spp_id', $id)->get();
+        if ($siswa->count() <= 0) {
+            $spp = Spp::find($id);
+            if ($spp) {
+                $spp->delete();
+                $this->emit('toastify', ['success','Berhasil Menghapus Spp']);
+            } else {
+                $this->emit('toastify', ['success','Berhasil Menghapus Spp']);
+            }
         } else {
-            $this->emit('toastify', ['success','Berhasil Menghapus Spp']);
+            $this->emit('toastify', ['danger', 'Tidak Bisa Menghapus Data Spp', 3000]);
+            $this->emit('toastify', ['danger', 'Terdapat Siswa yang terhubung pada data Spp', 3000]);
+            $this->emit('swal', ['warning', 'Disarankan untuk mengedit data', 2000]);
         }
     }
